parse config string during initial processing
