+++ src/helpers.h
@@ -542,6 +542,4 @@ std::wstring Scroller(const std::wstring &str, size_t &pos, size_t width);
 void writeCyclicBuffer(const NC::WBuffer &buf, NC::Window &w, size_t &start_pos,
                        size_t width, const std::wstring &separator);
 
-std::vector<int> parse_ratio(const std::string &s);
-
 #endif // NCMPCPP_HELPERS_H
