+++ src/utility/option_parser.h
@@ -38,8 +38,8 @@
 #include <boost/lexical_cast.hpp>
 #include <cassert>
 #include <stdexcept>
-#include <string>
 #include <unordered_map>
+#include <vector>
 
 [[noreturn]] inline void invalid_value(const std::string &v)
 {
@@ -76,9 +76,10 @@ std::vector<ValueT> list_of(const std::string &v)
 
 bool yes_no(const std::string &v);
 
+std::vector<int> parse_ratio(const std::string &s);
 bool integers_in_range(const std::string &v);
-std::string check_ratio_two(const std::string &v);
-std::string check_ratio_three(const std::string &v);
+std::vector<int> check_ratio_two(const std::string &v);
+std::vector<int> check_ratio_three(const std::string &v);
 
 ////////////////////////////////////////////////////////////////////////////////
